MOVE QUERY CLASSES TO THEIR OWN FOLDER
- in quizzes/graphql/query.py
This separation of concerns will make it easier to read and add new queries later down the road
